Refactor Rendering logic in ui file + consistent UI for confirmation prompt
